Moving IRange to Ranges
