+++ tests/Tests/Ranges/Adapters/IRange.cs
@@ -1,5 +1,5 @@
 using With;
-
+using With.Ranges;
 namespace Tests.Ranges.Adapters
 {
     /// <summary>
